Allow creating simple one-color images (#2)
# Summary
We want to be able to customize the background color for overlay operations. Currently it defaults to a transparent color. In order to allow this, we need to add the ability to create images of a single color.

This PR also bumps Rust / Elixir / deps and moves the `.cargo/config` file to satisfy Rustler.

Co-authored-by: Dana Daldoss <danadaldoss@gmail.com>